+++ test/unit/commands/ping.js
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 const assert = require('assert');
 
-const ping = require('../../commands/ping.js');
+const ping = require('../../../commands/ping.js');
 
 describe('ping', () => {
   it('returns pong', async () => {
